v1 participant endpoint uses induction records
- there was another endpoint that was run in parallel with existing
  fetch participant endpoint that was using induction records
- the old endpoint has been swapped out for this parallel endpoint
