+++ app/controllers/api/v1/ecf_participants_controller.rb
@@ -9,16 +9,16 @@ module Api
       include ApiPagination
       include ApiCsv
       include ApiFilter
-      include Api::ParticipantActions
+      include ParticipantActions
 
       def index
         respond_to do |format|
           format.json do
-            participant_hash = ParticipantSerializer.new(paginate(participants), { params: { mentor_ids: mentor_ids } }).serializable_hash
+            participant_hash = ParticipantFromInductionRecordSerializer.new(paginate(induction_records)).serializable_hash
             render json: participant_hash.to_json
           end
           format.csv do
-            participant_hash = ParticipantSerializer.new(participants, { params: { mentor_ids: mentor_ids } }).serializable_hash
+            participant_hash = ParticipantFromInductionRecordSerializer.new(induction_records).serializable_hash
             render body: to_csv(participant_hash)
           end
         end
@@ -27,68 +27,79 @@ module Api
     private
 
       def serialized_response(profile)
-        ParticipantSerializer
-          .new(profile)
+        relevant_induction_record = profile.relevant_induction_record(lead_provider: lead_provider)
+
+        ParticipantFromInductionRecordSerializer
+          .new(relevant_induction_record)
           .serializable_hash.to_json
       end
 
       def access_scope
-        LeadProviderApiToken.joins(cpd_lead_provider: [:lead_provider])
+        LeadProviderApiToken
+          .joins(cpd_lead_provider: [:lead_provider])
       end
 
       def lead_provider
         current_user.lead_provider
       end
 
-      def participants
-        participant_profiles = ParticipantProfile::ECF.where(id: participant_profile_ids)
-                                                      .joins(:participant_identity, :user)
-                                                      .includes(
-                                                        :participant_identity,
-                                                        :user,
-                                                        :cohort,
-                                                        :school,
-                                                        :ecf_participant_eligibility,
-                                                        :ecf_participant_validation_data,
-                                                        :schedule,
-                                                        :teacher_profile,
-                                                      )
+      def induction_records
+        scope = InductionRecord
+          .where(id: induction_record_ids_with_deduped_profiles)
+          .joins(participant_profile: { school_cohort: [:cohort] })
+          .where(participant_profile: { school_cohorts: { cohort: Cohort.current } })
+          .includes(participant_profile: [
+            :participant_identity,
+            :user,
+            :cohort,
+            :school,
+            :ecf_participant_eligibility,
+            :ecf_participant_validation_data,
+            :schedule,
+            :teacher_profile,
+            { mentor_profile: [:mentor] },
+          ])
 
         if updated_since.present?
-          participant_profiles = participant_profiles.where(user: { updated_at: updated_since.. })
-                                                     .order("\"user\".updated_at, \"user\".id")
+          scope = scope
+            .where(users: { updated_at: updated_since.. })
+            .order("users.updated_at, users.id")
         end
 
-        participant_profiles.order("\"user\".created_at")
+        scope.order("users.created_at")
       end
 
-      def participant_profile_ids
-        @participant_profile_ids ||= fetch_participant_profile_ids
-      end
-
-      def mentor_ids
-        # I can't figure out a way to preload the mentor IDs with the other data here, since that would include two copies of the
-        # users table, and active_record picks the wrong one. This horrible hack is to keep the number of db queries sane by
-        # constructing a hash from profile ID to mentor ID in a single query
-        ParticipantProfile::ECT.where(id: participant_profile_ids)
-                               .joins(:mentor)
-                               .pluck(:id, User.arel_table["id"])
-                               .to_h
-      end
-
-      def fetch_participant_profile_ids
-        # this retrieves the list of ECF participant profiles for the LeadProvider, one per
-        # teacher_profile (now participant_identity).
-        # Withdrawn profiles are included unless there is also an active one.
-        # The DISTINCT ON clause chooses the first record after ordering by status to do this.
-        inner_query = lead_provider
-          .ecf_participant_profiles
-          .select("DISTINCT ON (participant_profiles.participant_identity_id) participant_identity_id, participant_profiles.status, participant_profiles.id AS id")
-          .joins(:school_cohort)
-          .where(school_cohort: { cohort_id: Cohort.current.id })
-          .order(:participant_identity_id, status: :asc)
+      # inner most query
+      # this query deals with the following scenario
+      # given one profile with 2 induction records
+      # we only want to return the latest induction record
+      # we also exclude any partnerships that have been challenged
+      def induction_record_ids_with_deduped_induction_records
+        query = InductionRecord
+          .joins(induction_programme: { partnership: :lead_provider })
+          .joins(participant_profile: %i[user participant_identity])
+          .select("DISTINCT ON (participant_profiles.id) participant_profile_id, induction_records.id")
+          .where(induction_programme: { partnerships: { lead_provider: lead_provider } })
+          .where(induction_programme: { partnerships: { challenged_at: nil, challenge_reason: nil } })
+          .order("participant_profiles.id", start_date: :desc)
           .to_sql
-        ActiveRecord::Base.connection.query_values("SELECT id FROM (#{inner_query}) AS inner_query")
+
+        ActiveRecord::Base.connection.query_values("SELECT id FROM (#{query}) AS inner_query")
+      end
+
+      # second inner most query
+      # this query deals with where a user can have multiple profiles
+      # the work to map one user to one profile has not been done
+      # therefore we must work around and select correct profile
+      def induction_record_ids_with_deduped_profiles
+        query = InductionRecord
+          .where(id: induction_record_ids_with_deduped_induction_records)
+          .joins(participant_profile: [:participant_identity])
+          .select("DISTINCT ON (participant_profiles.participant_identity_id) participant_identity_id, induction_records.training_status, participant_profiles.created_at, induction_records.id")
+          .order("participant_profiles.participant_identity_id", "induction_records.training_status ASC", "participant_profiles.created_at DESC")
+          .to_sql
+
+        ActiveRecord::Base.connection.query_values("SELECT id FROM (#{query}) AS inner_query")
       end
     end
   end
