+++ spec/requests/api/v1/participants_spec.rb
@@ -15,18 +15,56 @@ RSpec.describe "Participants API", :with_default_schdules, type: :request do
     let(:bearer_token) { "Bearer #{token}" }
 
     before :each do
-      mentor_profile = create(:mentor_participant_profile, school_cohort: school_cohort)
-      create_list :ect_participant_profile, 2, mentor_profile: mentor_profile, school_cohort: school_cohort
+      mentor_profile = create(
+        :mentor_participant_profile,
+        school_cohort: school_cohort,
+      ).tap do |profile|
+        Induction::Enrol.call(participant_profile: profile, induction_programme: induction_programme)
+      end
+
+      profiles = create_list :ect_participant_profile, 2, mentor_profile: mentor_profile, school_cohort: school_cohort
+
+      profiles.each do |profile|
+        Induction::Enrol.call(participant_profile: profile, induction_programme: induction_programme)
+      end
+
       ect_teacher_profile_with_one_active_and_one_withdrawn_profile_record = ParticipantProfile::ECT.first.teacher_profile
-      create(:ect_participant_profile,
-             :withdrawn_record,
-             teacher_profile: ect_teacher_profile_with_one_active_and_one_withdrawn_profile_record,
-             school_cohort: school_cohort)
+
+      create(
+        :ect_participant_profile,
+        :withdrawn_record,
+        teacher_profile: ect_teacher_profile_with_one_active_and_one_withdrawn_profile_record,
+        school_cohort: school_cohort,
+      ).tap do |profile|
+        Induction::Enrol.call(participant_profile: profile, induction_programme: induction_programme).tap do |induction_record|
+          induction_record.update!(training_status: "withdrawn", induction_status: "withdrawn")
+        end
+      end
     end
 
-    let!(:withdrawn_ect_profile_record) { create(:ect_participant_profile, :withdrawn_record, school_cohort: school_cohort) }
+    let!(:withdrawn_ect_profile_record) do
+      create(
+        :ect_participant_profile,
+        :withdrawn_record,
+        school_cohort: school_cohort,
+      ).tap do |profile|
+        Induction::Enrol.call(participant_profile: profile, induction_programme: induction_programme).tap do |induction_record|
+          induction_record.update!(training_status: "withdrawn", induction_status: "withdrawn")
+        end
+      end
+    end
+
     let(:user) { create(:user) }
-    let(:early_career_teacher_profile) { create(:ect_participant_profile, school_cohort: school_cohort, user: user) }
+
+    let(:early_career_teacher_profile) do
+      create(
+        :ect_participant_profile,
+        school_cohort: school_cohort,
+        user: user,
+      ).tap do |profile|
+        Induction::Enrol.call(participant_profile: profile, induction_programme: induction_programme)
+      end
+    end
 
     context "when authorized" do
       before do
@@ -231,7 +269,7 @@ RSpec.describe "Participants API", :with_default_schdules, type: :request do
           expect(withdrawn_record_row["eligible_for_funding"]).to be_empty
           expect(withdrawn_record_row["pupil_premium_uplift"]).to eql(withdrawn_ect_profile_record.pupil_premium_uplift.to_s)
           expect(withdrawn_record_row["sparsity_uplift"]).to eql(withdrawn_ect_profile_record.sparsity_uplift.to_s)
-          expect(withdrawn_record_row["training_status"]).to eql(withdrawn_ect_profile_record.training_status)
+          expect(withdrawn_record_row["training_status"]).to eql(withdrawn_ect_profile_record.induction_records.first.training_status)
         end
 
         it "ignores pagination parameters" do
@@ -264,20 +302,13 @@ RSpec.describe "Participants API", :with_default_schdules, type: :request do
         end
       end
 
-      it_behaves_like "JSON Participant Change schedule endpoint" do
-        let!(:induction_record) do
-          Induction::Enrol.call(participant_profile: early_career_teacher_profile, induction_programme: induction_programme)
-        end
-      end
+      it_behaves_like "JSON Participant Change schedule endpoint"
 
       it_behaves_like "JSON Participant Deferral endpoint", "participant" do
         let(:url)               { "/api/v1/participants/#{early_career_teacher_profile.user.id}/defer" }
         let(:withdrawal_url)    { "/api/v1/participants/#{early_career_teacher_profile.user.id}/withdraw" }
         let(:params)            { { data: { attributes: { course_identifier: "ecf-induction", reason: "career-break" } } } }
         let(:withdrawal_params) { { data: { attributes: { course_identifier: "ecf-induction", reason: "left-teaching-profession" } } } }
-        let!(:induction_record) do
-          Induction::Enrol.call(participant_profile: early_career_teacher_profile, induction_programme: induction_programme)
-        end
       end
 
       it_behaves_like "JSON Participant Resume endpoint", "participant" do
@@ -285,9 +316,7 @@ RSpec.describe "Participants API", :with_default_schdules, type: :request do
         let(:withdrawal_url)    { "/api/v1/participants/#{early_career_teacher_profile.user.id}/withdraw" }
         let(:params)            { { data: { attributes: { course_identifier: "ecf-induction" } } } }
         let(:withdrawal_params) { { data: { attributes: { course_identifier: "ecf-induction", reason: "left-teaching-profession" } } } }
-        let!(:induction_record) do
-          Induction::Enrol.call(participant_profile: early_career_teacher_profile, induction_programme: induction_programme)
-        end
+
         before do
           put "/api/v1/participants/#{early_career_teacher_profile.user.id}/defer",
               params: { data: { attributes: { course_identifier: "ecf-induction", reason: "career-break" } } }
